requirements.txt
# Core AI / Agents
crewai==1.7.2
crewai-tools==1.7.2
langchain-openai==0.3.23
python-dotenv==1.1.1

# API & Server
fastapi==0.110.3
uvicorn[standard]==0.31.1
pydantic==2.11.9
pydantic-settings==2.10.1

# Document Processing
pypdf2==3.0.1
python-multipart==0.0.9

# Database & Storage
sqlalchemy==2.0.29
asyncpg==0.29.0
boto3==1.34.122

# Utilities
loguru==0.7.2
